// src/types/AggregationBuilder/aggregation-stages.ts
export enum AggregationStages {
    AddFields = "$addFields",
    Bucket = "$bucket",
    BucketAuto = "$bucketAuto",
    ChangeStream = "$changeStream",
    CollStats = "$collStats",
    Count = "$count",
    CurrentOp = "$currentOp",
    Densify = "$densify",
    Documents = "$documents",
    Facet = "$facet",
    Fill = "$fill",
    GeoNear = "$geoNear",
    GraphLookup = "$graphLookup",
    Group = "$group",
    IndexStats = "$indexStats",
    Limit = "$limit",
    Lookup = "$lookup",
    Match = "$match",
    Merge = "$merge",
    Out = "$out",
    PlanCacheStats = "$planCacheStats",
    Project = "$project",
    Redact = "$redact",
    ReplaceRoot = "$replaceRoot",
    ReplaceWith = "$replaceWith",
    Sample = "$sample",
    Search = "$search",
    SearchMeta = "$searchMeta",
    Set = "$set",
    SetWindowFields = "$setWindowFields",
    ShardedDataDistribution = "$shardedDataDistribution",
    Skip = "$skip",
    Sort = "$sort",
    SortByCount = "$sortByCount",
    UnionWith = "$unionWith",
    Unset = "$unset",
    Unwind = "$unwind"
}